requirements.txt
appdirs==1.4.4
asgiref==3.3.1
click==7.1.2
dataclasses==0.8
decorator==4.4.2
Django==3.1.3
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
image==1.5.33
llvmlite==0.34.0
Mako==1.1.3
MarkupSafe==1.1.1
multitasking==0.0.9
numba==0.51.2
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pkg-resources==0.0.0
pycuda==2020.1
pytools==2020.4.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
